Move assessment containers
