+++ src/containers/assessment/index.ts
@@ -1,56 +1,38 @@
 import { connect, MapDispatchToProps, MapStateToProps } from 'react-redux'
+import { withRouter } from 'react-router'
 import { bindActionCreators, Dispatch } from 'redux'
 
-import {
-  changeActiveTab,
-  changeEditorWidth,
-  changeSideContentHeight,
-  chapterSelect,
-  clearReplOutput,
-  evalEditor,
-  evalRepl,
-  fetchAssessment,
-  handleInterruptExecution,
-  updateEditorValue,
-  updateReplValue,
-  WorkspaceLocation
-} from '../../actions'
-import Assessment, { DispatchProps, OwnProps, StateProps } from '../../components/assessment'
+import { fetchAssessmentOverviews } from '../../actions/session'
+import { resetAssessmentWorkspace, updateCurrentAssessmentId } from '../../actions/workspaces'
+import AssessmentListing, {
+  IDispatchProps,
+  IOwnProps,
+  IStateProps
+} from '../../components/assessment/AssessmentListing'
+import { IAssessmentOverview } from '../../components/assessment/assessmentShape'
 import { IState } from '../../reducers/states'
 
-const mapStateToProps: MapStateToProps<StateProps, OwnProps, IState> = (state, props) => {
-  return {
-    assessment: state.session.assessments.get(props.assessmentId),
-    editorValue: state.workspaces.assessment.editorValue,
-    isRunning: state.workspaces.assessment.isRunning,
-    activeTab: state.workspaces.assessment.sideContentActiveTab,
-    editorWidth: state.workspaces.assessment.editorWidth,
-    sideContentHeight: state.workspaces.assessment.sideContentHeight,
-    output: state.workspaces.assessment.output,
-    replValue: state.workspaces.assessment.replValue
+const mapStateToProps: MapStateToProps<IStateProps, IOwnProps, IState> = (state, props) => {
+  const categoryFilter = (overview: IAssessmentOverview) =>
+    overview.category === props.assessmentCategory
+  const stateProps: IStateProps = {
+    assessmentOverviews: state.session.assessmentOverviews
+      ? state.session.assessmentOverviews.filter(categoryFilter)
+      : undefined,
+    storedAssessmentId: state.workspaces.currentAssessment,
+    storedQuestionId: state.workspaces.currentQuestion
   }
+  return stateProps
 }
 
-const location: WorkspaceLocation = 'assessment'
-
-const mapDispatchToProps: MapDispatchToProps<DispatchProps, {}> = (dispatch: Dispatch<any>) =>
-  bindActionCreators<DispatchProps>(
+const mapDispatchToProps: MapDispatchToProps<IDispatchProps, {}> = (dispatch: Dispatch<any>) =>
+  bindActionCreators(
     {
-      handleAssessmentFetch: fetchAssessment,
-      handleChangeActiveTab: (activeTab: number) => changeActiveTab(activeTab, location),
-      handleChapterSelect: (chapter: any, changeEvent: any) =>
-        chapterSelect(chapter, changeEvent, location),
-      handleEditorEval: () => evalEditor(location),
-      handleEditorValueChange: (val: string) => updateEditorValue(val, location),
-      handleEditorWidthChange: (widthChange: number) => changeEditorWidth(widthChange, location),
-      handleInterruptEval: () => handleInterruptExecution(location),
-      handleReplEval: () => evalRepl(location),
-      handleReplOutputClear: () => clearReplOutput(location),
-      handleReplValueChange: (newValue: string) => updateReplValue(newValue, location),
-      handleSideContentHeightChange: (heightChange: number) =>
-        changeSideContentHeight(heightChange, location)
+      handleAssessmentOverviewFetch: fetchAssessmentOverviews,
+      handleResetAssessmentWorkspace: resetAssessmentWorkspace,
+      handleUpdateCurrentAssessmentId: updateCurrentAssessmentId
     },
     dispatch
   )
 
-export default connect(mapStateToProps, mapDispatchToProps)(Assessment)
+export default withRouter(connect(mapStateToProps, mapDispatchToProps)(AssessmentListing))
